04_Deployement/consumer/requirements.txt
numpy<2
pandas==2.1.4
scikit-learn==1.4.2
xgboost==1.7.6
mlflow==2.21.0
psycopg2-binary==2.9.9
confluent-kafka==2.3.0
python-dotenv==1.0.1
sqlalchemy==2.0.30
boto3==1.34.101
fsspec==2024.3.1
s3fs==2024.3.1
